restructured for pypi submission
